Change Getting Started to use RSpec
For now, using `--test=rspec`, but we anticipate to change RSpec to be
the default testing framework for Hanami.

MiniTest is simpler, but it's less familiar to most people.
Additionally, when this was created, Hanami itself used MiniTest.
We've since switched over to RSpec, since we needed some of its more
powerful features (and it's what most of use most of the time).